fix(linear): correct Project sync issues from PR #2022
- Progress field: int → float64 (Linear returns 0.0-1.0)
- CreateProject: teamId → teamIds (projects span multiple teams)
- ProjectToEpic: map completedAt → ClosedAt, set CloseReason for canceled
- ProjectToEpic: canonicalize ExternalRef URL, remove raw Linear UUID as ID
- MapEpicToProjectState: map StatusInProgress → "started" for correct round-trip

+++ internal/linear/types.go
@@ -175,7 +175,7 @@ type Project struct {
 	SlugId      string `json:"slugId"`
 	URL         string `json:"url"`
 	State       string `json:"state"` // "planned", "started", "paused", "completed", "canceled"
-	Progress    int    `json:"progress"`
+	Progress    float64 `json:"progress"`
 	CreatedAt   string `json:"createdAt"`
 	UpdatedAt   string `json:"updatedAt"`
 	CompletedAt string `json:"completedAt,omitempty"`
